+++ ios/ads/THEOplayerRCTSourceDescriptionBuilder+Ads.swift
@@ -56,7 +56,7 @@ extension THEOplayerRCTSourceDescriptionBuilder {
             case "google-ima":
                 return THEOplayerRCTSourceDescriptionBuilder.buildSingleGoogleIMAAdsDescription(adsData)
             case "theoads":
-                return THEOplayerRCTSourceDescriptionBuilder.buildSingleTheoAdsDescription(adsData)
+                return THEOplayerRCTSourceDescriptionBuilder.buildSingleTHEOadsDescription(adsData)
             default:
                 if DEBUG_SOURCE_DESCRIPTION_BUIDER  { PrintUtils.printLog(logText: "[NATIVE] We currently require and only support the 'google-ima' or 'sgai' integration in the 'ads' description.") }
             }